feat - reorg, switch from getOrgUsers to searchOrgUsers (#1044)
* feat - reorg, switch from getOrgUsers to searchOrgUsers

* fixed search string err
