+++ wondrous-app/components/organization/analytics/index.tsx
@@ -4,7 +4,7 @@ import { useInView } from 'react-intersection-observer';
 import AdapterDateFns from '@mui/lab/AdapterDateFns';
 import LocalizationProvider from '@mui/lab/LocalizationProvider';
 import { format } from 'date-fns';
-import { GET_AUTOCOMPLETE_USERS, GET_COMPLETED_TASKS_BETWEEN_TIME_PERIOD, GET_ORG_USERS } from 'graphql/queries';
+import { GET_AUTOCOMPLETE_USERS, GET_COMPLETED_TASKS_BETWEEN_TIME_PERIOD, SEARCH_ORG_USERS } from 'graphql/queries';
 import palette from 'theme/palette';
 import { SafeImage } from 'components/Common/Image';
 import DefaultUserImage from 'components/Common/Image/DefaultUserImage';
@@ -420,22 +420,21 @@ export const getContributorTaskData = (data) => {
 function Analytics(props) {
   const { orgData = {} } = props;
   const { id: orgId } = orgData;
-  const [ref, inView] = useInView({});
   const [assignee, setAssignee] = useState(null);
   const [payoutModal, setPayoutModal] = useState(false);
   const [assigneeString, setAssigneeString] = useState('');
-  const [getAutocompleteUsers, { data: autocompleteData }] = useLazyQuery(GET_AUTOCOMPLETE_USERS);
-  const [getOrgUsers, { data: orgUsersData }] = useLazyQuery(GET_ORG_USERS, {});
+  const [searchOrgUsers, { data: orgUsersData }] = useLazyQuery(SEARCH_ORG_USERS, {});
 
   useEffect(() => {
     if (orgId) {
-      getOrgUsers({
+      searchOrgUsers({
         variables: {
           orgId,
+          searchString: '',
         },
       });
     }
-  }, [orgId, getOrgUsers]);
+  }, [orgId, searchOrgUsers]);
 
   const today = new Date();
   const tomorrow = new Date(today.getFullYear(), today.getMonth(), today.getDate() + 1);
@@ -467,6 +466,15 @@ function Analytics(props) {
     }
   }, [orgId, fromTime, toTime, assignee?.value]);
 
+  const handleInputChange = (event, newInputValue) => {
+    setAssigneeString(newInputValue);
+    searchOrgUsers({
+      variables: {
+        searchString: newInputValue,
+        orgId,
+      },
+    });
+  };
   return (
     <Wrapper orgData={orgData}>
       <PayoutModal
@@ -500,11 +508,7 @@ function Analytics(props) {
         </LocalizationProvider>
         <HeaderText>by</HeaderText>
         <StyledAutocompletePopper
-          options={
-            assigneeString
-              ? filterUsers(autocompleteData?.getAutocompleteUsers)
-              : filterOrgUsers(orgUsersData?.getOrgUsers)
-          }
+          options={filterUsers(orgUsersData?.searchOrgUsers)}
           onOpen={() => {}}
           renderInput={(params) => {
             const InputProps = {
@@ -544,14 +548,7 @@ function Analytics(props) {
           }}
           value={assignee}
           inputValue={assigneeString}
-          onInputChange={(event, newInputValue) => {
-            setAssigneeString(newInputValue);
-            getAutocompleteUsers({
-              variables: {
-                username: newInputValue,
-              },
-            });
-          }}
+          onInputChange={handleInputChange}
           onChange={(_, __, reason) => {
             if (reason === 'clear') {
               setAssignee(null);
